Ignore serial and update zone file only if necessary
Update a metadata file containing mock data that will be reviewed before
a zone is updated to determine if the content is identical.

Additionally, switch from a time-based serial number to an
iteration-based serial number that is stored in the metadata file
aswell.

This commit will require manual-intervention as the serial is unlikely
to be larger than the current date.

Fix typo: renamed updatezone.yaml => updatezone.yml